Homework #05. Task01. Review code. Task02. Review code.
