+++ src/main/java/lec05/task01/classes/IoSort.java
@@ -1,0 +1,37 @@
+package lec05.task01.classes;
+
+import lec05.task01.data.StatFinder;
+import lec05.task01.interfaces.FileSorter;
+import lombok.NonNull;
+import java.io.*;
+import java.util.*;
+import java.util.stream.Stream;
+
+//@UtilityClass
+public class IoSort implements FileSorter {
+
+    public StatFinder sort(@NonNull String inputFileName, @NonNull String outputFileName) throws IOException {
+
+        int foundWords = 0;
+        HashSet<String> words;
+
+        WordFinder wordFinder = new WordFinder();
+        words = wordFinder.find(inputFileName);
+
+        foundWords = words.size();
+
+        try (FileWriter writer = new FileWriter(outputFileName);
+             BufferedWriter bw = new BufferedWriter(writer)) {
+            Stream<String> stream = words.stream().sorted(Comparator.naturalOrder());
+            stream.forEach((s) -> {
+                try {
+                    bw.write(s);
+                    bw.newLine();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            });
+        }
+        return new StatFinder(foundWords, words.size());
+    }
+}
