+++ src/main/java/lec05/task01/classes/NioSort.java
@@ -1,15 +1,19 @@
-package lec05.task01;
+package lec05.task01.classes;
 
 import lec05.task01.data.CharacterTables;
 import lec05.task01.data.StatFinder;
 import lec05.task01.interfaces.FileSorter;
 import lombok.NonNull;
 import java.io.*;
-import java.util.*;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.HashSet;
 import java.util.stream.Stream;
 
 //@UtilityClass
-public class IoSort implements FileSorter {
+public class NioSort implements FileSorter {
 
     private boolean isLetter(char ch) {
         return Arrays.binarySearch(CharacterTables.latinLettersCharArray, ch) >= 0 ||
@@ -21,12 +25,11 @@ public class IoSort implements FileSorter {
         int foundWords = 0;
         HashSet<String> words = new HashSet<>();
 
-        try (FileReader reader = new FileReader(inputFileName);
-             BufferedReader br = new BufferedReader(reader)) {
+        try (BufferedReader br = Files.newBufferedReader(Paths.get(inputFileName))) {
             StringBuilder str = new StringBuilder();
             int ch;
             while ((ch = br.read()) != -1) {
-                if (isLetter((char) ch)) {
+                if (isLetter((char)ch)) {
                     str.append((char) ch);
                 } else {
                     if (str.toString().isEmpty()) continue;
@@ -37,8 +40,7 @@ public class IoSort implements FileSorter {
             }
         }
 
-        try (FileWriter writer = new FileWriter(outputFileName);
-             BufferedWriter bw = new BufferedWriter(writer)) {
+        try (BufferedWriter bw = Files.newBufferedWriter(Paths.get(outputFileName))) {
             Stream<String> stream = words.stream().sorted(Comparator.naturalOrder());
             stream.forEach((s) -> {
                 try {
@@ -51,4 +53,5 @@ public class IoSort implements FileSorter {
         }
         return new StatFinder(foundWords, words.size());
     }
+
 }
